+++ TDTChocolate/FoundationAdditions/NSArray+TDTFunctionalAdditions.h
@@ -48,4 +48,11 @@
  */
 - (BOOL)tdt_anyWithBlock:(TDTPredicateBlock)predicate;
 
+/**
+ Construct a dictionary from (key, value) pairs.
+
+ @pre The reciever is expected to be of the form @p @[@[key, value] ..].
+ */
+- (NSDictionary *)tdt_dictionaryFromPairs;
+
 @end